+++ v2/AciduinoV2/src/sequencer/acid_sequencer.h
@@ -85,8 +85,8 @@ class AcidSequencerClass
       uint8_t getTrackChannel(uint8_t track);
       uint8_t getTrackPort(uint8_t track) { return 0; };
       uint8_t getTrackNumber();
-      uint8_t getTrackMaxLength(uint8_t track);
-      uint8_t getTrackLength(uint8_t track);
+      uint16_t getTrackMaxLength(uint8_t track);
+      uint16_t getTrackLength(uint8_t track);
       void setTrackLength(uint8_t track, uint16_t length);
       void clearTrack(uint8_t track);
       void acidRandomize(uint8_t track, uint8_t fill, uint8_t param_1 = 0, uint8_t param_2 = 0, uint8_t param_3 = 0, uint8_t param_4 = 0, uint8_t param_5 = 0, uint8_t param_6 = 0);
