+++ v2/AciduinoV2/src/sequencer/engine_808.h
@@ -120,7 +120,7 @@ class Engine808 : public Engine
       void setStepData(uint8_t track, uint8_t step, uint8_t data);
       uint8_t getStepData(uint8_t track, uint8_t step);
       uint8_t getCurrentStep(uint8_t track);
-      uint8_t getTrackLength(uint8_t track);
+      uint16_t getTrackLength(uint8_t track);
       uint8_t getTrackChannel(uint8_t track);
       void setShiftPos(uint8_t track, int8_t shift);
       int8_t getShiftPos(uint8_t track);
